+++ zigbee-map-app/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-02-19
+### Added
+- Add "Done" button in the Hubitat app - `@dnickel`
+- Click the address of any device in the "Devices" tab to add it to the Interview Queue - `@hubitrep`
+- Use relative URL when opening the HTML app - `@jlv`
+- Mark devices that failed the Interview  - `@kahn-hubitat`
+- Show Interview Queue size
+- Show "duplex" links by default
 
 ## [1.0.0] - 2024-02-16
 ### Added